backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
httpx
python-jose[cryptography]